pyproject.toml
[tool.poetry]
name = "tab-scraper"
version = "0.1.0"
description = "Scrape TAB data for betting information and analyse results"
authors = ["Kieran <kieranhitchcock28@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
pymongo = "^4.10.1"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
